+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Fix: inline images incorrectly always created a new line in DOCX ➡️ HTML converter
 - Preserve hyperlinks on images in DOCX ➡️ HTML converter
 - Fix: hyperlinks and images in header/footer/footnote/endnote were sometimes lost in DOCX converters
+- Preserve image alternate text in DOCX ➡️ HTML/Markdown converter
+- Optionally write alternate text (if available) in place of images in DOCX ➡️ plain text converter
 
 ## 0.18.1 - 2026.04.10